Should I add .idea/caches/build_file_checksums.ser to .gitignore?
All files in your .idea folders are related to your IDE (Android studio in this case). And as you said this files is created at the opening of the project. So you should add this file to your .gitignore because these files can contains some contents that is not necessary like your absolute path.